preliminary_contest/deep_learning/requirements.txt
absl-py==0.12.0
apturl==0.5.2
astor==0.8.1
bert4keras==0.10.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
compiler==0.2.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
filelock==3.0.12
fire==0.4.0
gast==0.4.0
gensim==3.8.3
glove-python-binary==0.2.0
google-pasta==0.2.0
grpcio==1.36.1
h5py==3.2.1
httplib2==0.14.0
idna==2.8
iterative-stratification==0.1.6
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.3.4
netifaces==0.10.4
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
packaging==20.9
pandas==1.2.3
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2021.3.17
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
smart-open==4.2.0
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
testresources==2.0.1
threadpoolctl==2.1.0
tokenizers==0.10.1
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.59.0
transformers==4.4.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
